add remove_call & getters from pool fn
