+++ sidechain/top-pool-executor/src/lib.rs
@@ -23,12 +23,50 @@ compile_error!("feature \"std\" and feature \"sgx\" cannot be enabled at the sam
 
 #[cfg(all(not(feature = "std"), feature = "sgx"))]
 extern crate sgx_tstd as std;
-
 // Re-export module to properly feature gate sgx and regular std environment.
 #[cfg(all(not(feature = "std"), feature = "sgx"))]
 pub mod sgx_reexport_prelude {
 	pub use thiserror_sgx as thiserror;
 }
 
+pub mod call_operator;
 pub mod error;
-pub mod top_pool_operation_executor;
+pub mod getter_operator;
+
+// Re-exports
+pub use call_operator::TopPoolCallOperator;
+pub use getter_operator::TopPoolGetterOperator;
+
+use itp_stf_executor::traits::{StfExecuteTimedCallsBatch, StfExecuteTimedGettersBatch};
+use itp_types::H256;
+use its_primitives::traits::{Block as SidechainBlockT, SignedBlock as SignedBlockT};
+use its_state::{SidechainState, SidechainSystemExt, StateHash};
+use its_top_pool_rpc_author::traits::{AuthorApi, OnBlockCreated, SendState};
+use sgx_externalities::SgxExternalitiesTrait;
+use sp_runtime::{traits::Block as BlockT, MultiSignature};
+use std::{marker::PhantomData, sync::Arc};
+
+/// Executes operations on the top pool
+///
+/// Operations can either be Getters or Calls
+pub struct TopPoolOperationHandler<PB, SB, RpcAuthor, StfExecutor> {
+	rpc_author: Arc<RpcAuthor>,
+	stf_executor: Arc<StfExecutor>,
+	_phantom: PhantomData<(PB, SB)>,
+}
+
+impl<PB, SB, RpcAuthor, StfExecutor> TopPoolOperationHandler<PB, SB, RpcAuthor, StfExecutor>
+where
+	PB: BlockT<Hash = H256>,
+	SB: SignedBlockT<Public = sp_core::ed25519::Public, Signature = MultiSignature>,
+	SB::Block: SidechainBlockT<ShardIdentifier = H256, Public = sp_core::ed25519::Public>,
+	RpcAuthor:
+		AuthorApi<H256, PB::Hash> + OnBlockCreated<Hash = PB::Hash> + SendState<Hash = PB::Hash>,
+	StfExecutor: StfExecuteTimedCallsBatch + StfExecuteTimedGettersBatch,
+	<StfExecutor as StfExecuteTimedCallsBatch>::Externalities:
+		SgxExternalitiesTrait + SidechainState + SidechainSystemExt + StateHash,
+{
+	pub fn new(rpc_author: Arc<RpcAuthor>, stf_executor: Arc<StfExecutor>) -> Self {
+		TopPoolOperationHandler { rpc_author, stf_executor, _phantom: Default::default() }
+	}
+}
